+++ src/common/sp-config.adb
@@ -17,7 +17,6 @@
 with Ada.Directories;
 with Ada.Strings.Unbounded;
 with Ada.Text_IO;
-with Dir_Iterators.Ancestor;
 with SP.File_System;
 with SP.Platform;
 
@@ -78,28 +77,10 @@ package body SP.Config is
         end;
     end Create_Local_Config;
 
-    -- Finds the config which is the closest ancestor to the given directory.
-    function Closest_Config (Dir_Name : String) return ASU.Unbounded_String with
-        Pre  => AD.Exists (Dir_Name),
-        Post => (Closest_Config'Result = ASU.Null_Unbounded_String)
-        or else FS.Is_File (ASU.To_String (Closest_Config'Result))
-    is
-        Ancestors  : constant Dir_Iterators.Ancestor.Ancestor_Dir_Walk := Dir_Iterators.Ancestor.Walk (Dir_Name);
-        Next_Trial : ASU.Unbounded_String;
-    begin
-        for Ancestor of Ancestors loop
-            --  Look for hidden ".septum/" directories in the path.
-            Next_Trial := ASU.To_Unbounded_String (Ancestor & "/" & Local_Config_Dir_Name & "/" & Config_File_Name);
-            if FS.Is_File (ASU.To_String (Next_Trial)) then
-                return Next_Trial;
-            end if;
-        end loop;
-        return ASU.Null_Unbounded_String;
-    end Closest_Config;
-
     function Config_Locations return String_Sets.Set is
         Config_Dir : constant SP.Strings.String_Holders.Holder := SP.Platform.Config_Dir;
-        Current_Dir_Config : constant ASU.Unbounded_String := Closest_Config (Ada.Directories.Current_Directory);
+        Current_Dir_Config : constant ASU.Unbounded_String := ASU.To_Unbounded_String (
+            Ada.Directories.Current_Directory & "/" & Local_Config_Dir_Name & "/" & Config_File_Name);
     begin
         return Result : String_Sets.Set do
             --  Look for a global user config.